adding the "flip" effect to the countdown timer | refactored
