+++ app/globals.css
@@ -209,7 +209,6 @@
 }
 
 /* styles for the flip card*/
-/* Note, the developer knows little of the code below*/
 
 /* Flip Clock Styles */
 .upper-card,
